Move to Speed.hpp add additional functions
