+++ src/OpenLoco/Types.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <cstdint>
-#include <type_traits>
 
 namespace OpenLoco
 {
@@ -140,90 +139,5 @@ namespace OpenLoco
     namespace ZoomLevels
     {
         constexpr uint8_t max = 4;
-    }
-
-    template<typename Value>
-    struct SpeedTemplate
-    {
-        Value value;
-        constexpr Value getRaw() const { return value; }
-        constexpr explicit SpeedTemplate(Value val)
-            : value(val)
-        {
-        }
-
-        // Conversion function (only valid to more accuracy)
-        template<typename T>
-        constexpr operator SpeedTemplate<T>() const
-        {
-            static_assert(sizeof(T) > sizeof(Value));
-            constexpr auto shift = (sizeof(T) - sizeof(Value)) * 8;
-            return SpeedTemplate<T>(getRaw() << shift);
-        }
-
-        // Best accuracy between ThisType and OtherType
-        template<typename OtherType>
-        using BestType = std::conditional_t<(sizeof(Value) > sizeof(OtherType)), SpeedTemplate<Value>, SpeedTemplate<OtherType>>;
-
-        template<typename RhsT>
-        constexpr bool operator==(SpeedTemplate<RhsT> const& rhs)
-        {
-            return BestType<RhsT>(*this).value == BestType<RhsT>(rhs).value;
-        }
-        template<typename RhsT>
-        constexpr bool operator!=(SpeedTemplate<RhsT> const& rhs)
-        {
-            return !(*this == rhs);
-        }
-        template<typename RhsT>
-        constexpr bool operator>(SpeedTemplate<RhsT> const& rhs)
-        {
-            return BestType<RhsT>(*this).value > BestType<RhsT>(rhs).value;
-        }
-        template<typename RhsT>
-        constexpr bool operator<=(SpeedTemplate<RhsT> const& rhs)
-        {
-            return !(*this > rhs);
-        }
-        template<typename RhsT>
-        constexpr bool operator<(SpeedTemplate<RhsT> const& rhs)
-        {
-            return BestType<RhsT>(*this).value < BestType<RhsT>(rhs).value;
-        }
-        template<typename RhsT>
-        constexpr bool operator>=(SpeedTemplate<RhsT> const& rhs)
-        {
-            return !(*this < rhs);
-        }
-    };
-
-    using Speed32 = SpeedTemplate<int32_t>;
-    using Speed16 = SpeedTemplate<int16_t>;
-
-    constexpr auto speed16Null = Speed16(-1);
-    constexpr Speed16 toSpeed16(Speed32 speed)
-    {
-        return Speed16(speed.getRaw() / 65535);
-    }
-
-    namespace Literals
-    {
-        // Note: Only valid for 5 decimal places.
-        constexpr Speed32 operator"" _mph(long double speedMph)
-        {
-            uint16_t wholeNumber = speedMph;
-            uint64_t fraction = (speedMph - wholeNumber) * 100000;
-            return Speed32((static_cast<uint32_t>(wholeNumber) << 16) | static_cast<uint16_t>((fraction << 16) / 100000));
-        }
-
-        constexpr Speed16 operator""_mph(unsigned long long int speed)
-        {
-            return Speed16(speed);
-        }
-        static_assert(2.75_mph == Speed32(0x2C000));
-        static_assert(4.0_mph == Speed32(0x40000));
-        static_assert(14.0_mph == Speed32(0xE0000));
-        static_assert(0.333333_mph == Speed32(0x5555));
-        static_assert(2.0_mph == 2_mph);
     }
 }
