+++ src/OpenLoco/openloco.vcxproj
@@ -267,6 +267,7 @@
     <ClInclude Include="S5\SawyerStream.h" />
     <ClInclude Include="Scenario.h" />
     <ClInclude Include="ScenarioManager.h" />
+    <ClInclude Include="Speed.hpp" />
     <ClInclude Include="Station.h" />
     <ClInclude Include="StationManager.h" />
     <ClInclude Include="Things\Misc.h" />
